[CHORE] 구조 변경 (#125)
